pyproject.toml
[project]
name = "dados-google-genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.1.0",
    "pydantic-ai[logfire]>=0.0.23",
    "python-dotenv>=1.0.1",
]
